Move has3DTilesExtension into Scene
